+++ CHANGELOG.rst
@@ -5,8 +5,8 @@ Appendix C. Changelog
 :Author: Chris Warrick <chris@chriswarrick.com>
 :Copyright: © 2011-2018, Chris Warrick.
 :License: BSD (see /LICENSE or :doc:`Appendix B <LICENSE>`.)
-:Date: 2018-03-18
-:Version: 4.2.16
+:Date: 2018-03-25
+:Version: 4.2.17
 
 .. index:: CHANGELOG
 
@@ -46,6 +46,10 @@ More information can be found on GitHub in the `releases section <https://github
 
 Version History
 ===============
+
+4.2.17
+    * Add ``-G``, ``--get`` as ``-F``, ``--fetch`` alternatives (Issue #56)
+    * Pass ``pb -F`` to pacman’s file search option (Issue #56)
 
 4.2.16
     * Handle searches with spaces correctly (Issue #55)